+++ datatracer/__init__.py
@@ -4,13 +4,13 @@
 
 __author__ = 'MIT Data To AI Lab'
 __email__ = 'dailabmit@gmail.com'
-__version__ = '0.0.3'
+__version__ = '0.0.4.dev1'
 
 import os
 
 from mlblocks import discovery
 
-from datatracer.core import DataTracer
+from datatracer.core import PRETRAINED_DIR, DataTracer
 from datatracer.data import get_demo_data, load_dataset, load_datasets
 
 _BASE_PATH = os.path.abspath(os.path.dirname(__file__))
@@ -47,3 +47,17 @@ def get_primitives():
             List of the names of the available datatracer primitives.
     """
     return discovery.find_primitives('datatracer')
+
+
+def get_solvers():
+    """Get a list of the available datatracer solvers.
+
+    Returns:
+        list:
+            List of the names of the available datatracer solvers.
+    """
+    return [
+        pretrained_solver[:-3]
+        for pretrained_solver in os.listdir(PRETRAINED_DIR)
+        if pretrained_solver.endswith('.dt')
+    ]