+++ datatracer/column_map/__init__.py
@@ -1,7 +1,7 @@
-from datatracer.column_map.solver import ColumnMapSolver
-from datatracer.column_map.transformer import Transformer
+from datatracer.column_map.base import ColumnMapSolver
+from datatracer.column_map.basic import BasicColumnMapSolver
 
 __all__ = (
     'ColumnMapSolver',
-    'Transformer',
+    'BasicColumnMapSolver',
 )